feat: Add support for multiple payment apps and refactor AutoPay plugin
- Introduced SupportedApps enum to define available payment applications (GCASH, MAYA, METROBANK, BDO).
- Updated AutoPayPlugin interface to include PayMateOptions for startAutoPay method.
- Refactored AutoPayWeb class to streamline methods and remove unused ones.
- Implemented GatewayUrlProvider class in Android to manage gateway URLs for supported apps.
- Enhanced web.js and plugin.js files to reflect changes in the AutoPay implementation.
- Updated source maps and definitions to ensure consistency across the codebase.
