+++ android/src/main/java/com/bastion/inc/AppStateDetectorFactory.java
@@ -2,16 +2,17 @@ package com.bastion.inc;
 
 import android.content.Context;
 
+import com.bastion.inc.Enums.SupportedApp;
 import com.bastion.inc.Interfaces.AppStateDetector;
 import com.bastion.inc.StateDetectors.GCashStateDetector;
 import com.bastion.inc.StateDetectors.PayMayaStateDetector;
 
 public class AppStateDetectorFactory {
-    public static AppStateDetector getDetector(String appName, Context context){
+    public static AppStateDetector getDetector(SupportedApp appName, Context context){
         switch (appName){
-            case "PayMaya":
+            case MAYA:
                 return new PayMayaStateDetector(context);
-            case "GCash":
+            case GCASH:
                 return new GCashStateDetector(context);
             default:
                 throw new IllegalArgumentException("Unsupported app: " + appName);
