+++ android/src/main/java/com/bastion/inc/Enums/SupportedApp.java
@@ -1,0 +1,8 @@
+package com.bastion.inc.Enums;
+
+public enum SupportedApp {
+    MAYA,
+    GCASH,
+    METROBANK,
+    BDO
+}
