+++ src/web.ts
@@ -1,6 +1,6 @@
 import { WebPlugin } from '@capacitor/core';
 
-import type { AutoPayPlugin } from './definitions';
+import type { AutoPayPlugin, PayMateOptions } from './definitions';
 
 export class AutoPayWeb extends WebPlugin implements AutoPayPlugin {
   checkOverlayPermission(): Promise<any> {
@@ -51,26 +51,7 @@ export class AutoPayWeb extends WebPlugin implements AutoPayPlugin {
       }
     });
   }
-  stopNavigation(options: any): Promise<any> {
-    return options;
-  }
-  performGesture(options: { x: number; y: number; }): Promise<any> {
-    return new Promise((resolve, reject) => {
-      try{
-        resolve(options);
-      }catch(error){
-        reject(error);
-      }
-    });
-  }
-  startAutoPay(options: any): Promise<any>{
-    return options;
-  }
-  navigateGCash(options: any): Promise<any> {
-    return options;
-  }
-  async echo(options: { value: string }): Promise<{ value: string }> {
-    console.log('ECHO', options);
-    return options;
+  startAutoPay(options: PayMateOptions): Promise<any>{
+    return Promise.resolve(options);
   }
 }
